Move synapse/tests/common classes to synaspe/lib/testify
